[4-] Improve fluent syntax to have TypeSelectors to contain the Autofac's AS method call. (Instanciated<T>())
- remove NameMatching from SelfTypeSelector
