+++ src/Autofac.BindingByConvention/FluentSyntax/FluentTypeFilter.cs
@@ -12,25 +12,21 @@
     {
         private readonly IRegistrationBuilder<object, ScanningActivatorData, DynamicRegistrationStyle> builder;
 
+
         public FluentTypeFilter(
             IRegistrationBuilder<object, ScanningActivatorData, DynamicRegistrationStyle> builder,
+            FluentSelfTypeSelector fluentSelfTypeSelector,
             Func<Type, bool>[] predicates)
         {
             this.builder = builder;
+            this.AndTypesWillBe = fluentSelfTypeSelector;
 
             this.Except = new FluentExceptTypes(this, predicates);
         }
 
         protected IFluentExceptTypes Except { get; set; }
 
-        public void Instanciated<T>()
-            where T : RegistrationStrategyBase, new()
-        {
-            T instance = Activator.CreateInstance<T>();
 
-            this.builder.AsSelf();
-
-            instance.Apply(this.builder);            
-        }
+        public FluentSelfTypeSelector AndTypesWillBe { get; }
     }
 }
